+++ src/main/java/Thopz/AddEventCommand.java
@@ -1,20 +1,25 @@
+package Thopz;
+
 import java.io.IOException;
 
+
 /**
- * Adding a new todo task
+ * Adding a new event task
  */
-
-public class AddTODOCommand extends Command {
-
+public class AddEventCommand extends Command {
     private String desc;
+    private String start;
+    private String end;
 
-    public AddTODOCommand(String desc) {
-        this.desc = desc;
+    public AddEventCommand(String desc, String start, String end) {
+        this.desc =desc;
+        this.start = start;
+        this.end = end;
     }
 
     @Override
     public void perform (TaskList list, Storage storage, Ui ui) {
-        Todo task = new Todo(desc);
+        Events task = new Events(desc,start,end);
         list.addTask(task);
         try {
             storage.save(list.getAllTasks());
@@ -24,5 +29,4 @@ public class AddTODOCommand extends Command {
         }
         ui.showAddition(task,list.size());
     }
-
 }
