+++ src/main/java/Thopz/Command.java
@@ -1,4 +1,4 @@
-/**
+package Thopz; /**
  * Abstract command class
  */
