+++ src/main/java/Thopz/Events.java
@@ -1,10 +1,13 @@
+package Thopz;
+
 import java.time.DateTimeException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
 
-public class Deadline extends Task{
-    private String due;
+public class Events extends Task {
+
+    private String start;
+    private String end;
     public static final DateTimeFormatter[] INPUT_FORMATS = new DateTimeFormatter[] {
             DateTimeFormatter.ofPattern("dd/MM/yyyy HH:mm"),
             DateTimeFormatter.ofPattern("dd-MM-yyyy HH:mm"),
@@ -13,10 +16,17 @@ public class Deadline extends Task{
     };
     public static final DateTimeFormatter OUTPUT = DateTimeFormatter.ofPattern("MMM dd yyyy, HH:mm");
 
-    public Deadline (String tasks, String string) {
-        super(tasks);
-        this.due = parseDate(string.trim());
-        if(this.due==null) {
+
+    public Events(String desc, String start, String end) {
+        super(desc);
+        this.start = start;
+        this.end = end;
+        this.start = parseDate(start.trim());
+        if(this.start==null) {
+            throw new IllegalArgumentException("give in date and time format");
+        }
+        this.end = parseDate(end.trim());
+        if(this.end==null) {
             throw new IllegalArgumentException("give in date and time format");
         }
     }
@@ -33,15 +43,15 @@ public class Deadline extends Task{
         return dateString;
     }
 
-
-
     @Override
     public String saveformat() {
-        return  "D / " + (getTaskStatus() ? "1" : "0") + " / " + getTasks() + " / " + due;
+        return "E / " + (getTaskStatus() ? "1" : "0") + " / " + getTasks()
+                + " / " + start + " / " + end;
     }
 
     @Override
     public String toString() {
-        return "[D]" + super.toString() + " (due by "  + due +")";
+        return "[E]" + super.toString() + " (Starts on : " + start + ", ends by : " + end + ")";
     }
+
 }
