+++ src/main/java/Thopz/ExitCommand.java
@@ -1,4 +1,4 @@
-import java.io.IOException;
+package Thopz;
 
 /**
  * Stops the chatbot from running
