+++ src/main/java/Thopz/Parser.java
@@ -1,4 +1,4 @@
-import java.io.IOException;
+package Thopz;
 
 /**
  * Parser class to parse user inputs
