+++ src/main/java/Thopz/Task.java
@@ -1,3 +1,5 @@
+package Thopz;
+
 public class Task {
     public String tasks;
     private boolean done;
