+++ src/main/java/Thopz/Ui.java
@@ -1,3 +1,5 @@
+package Thopz;
+
 import java.util.List;
 import java.util.Scanner;
